--HG--
branch : issue4-osgi_bundles_conversion

+++ org.jenetics/src/main/java/org/jenetics/stat/Mean.java
@@ -87,7 +87,7 @@ public class Mean<N extends Number> extends MappedAccumulator<N> {
 			_mean = 0;
 		}
 
-		_mean += (value.doubleValue() - _mean)/(double)(++_samples);
+		_mean += (value.doubleValue() - _mean)/(++_samples);
 	}
 
 	@Override
