+++ org.jenetics/src/test/java/org/jenetics/stat/HistogramTest.java
@@ -36,6 +36,7 @@ import org.jenetics.util.RandomRegistry;
 
 /**
  * @author <a href="mailto:franz.wilhelmstoetter@gmx.at">Franz Wilhelmstötter</a>
+ * @version <em>$Date$</em>
  */
 public class HistogramTest
 	extends MappedAccumulatorTester<Histogram<Double>>
@@ -48,14 +49,14 @@ public class HistogramTest
 			final Random random = RandomRegistry.getRandom();
 			final double min = random.nextInt(100) + 100;
 			final double max = random.nextInt(100) + 100 + min;
-			
+
 			final Histogram<Double> histogram = Histogram.valueOf(
 						min, max, random.nextInt(10) + 5
 					);
 			for (int i = 0; i < 1000; ++i) {
 				histogram.accumulate(random.nextGaussian()*(max - min) + min);
 			}
-			
+
 			return histogram;
 		}
 	};
@@ -63,67 +64,67 @@ public class HistogramTest
 	protected Factory<Histogram<Double>> getFactory() {
 		return _factory;
 	}
-	
+
 	@Test
 	public void createDouble() {
 		final double begin = 12;
 		final double end = 123;
 		final int elements = 10;
-		
+
 		Histogram<Double> histogram = Histogram.valueOf(begin, end, elements);
 		Assert.assertEquals(histogram.length(), elements);
 		Assert.assertEquals(histogram.getHistogram(), new long[elements]);
 	}
-	
+
 	@Test
 	public void createFloat64() {
 		final Float64 begin = Float64.valueOf(12);
 		final Float64 end = Float64.valueOf(123);
 		final int elements = 10;
-		
+
 		Histogram<Float64> histogram = Histogram.valueOf(begin, end, elements);
 		Assert.assertEquals(histogram.length(), elements);
 		Assert.assertEquals(histogram.getHistogram(), new long[elements]);
 	}
-	
+
 	@Test
 	public void createLong() {
 		final long begin = 0;
 		final long end = 1000;
 		final int elements = 9;
-		
+
 		Histogram<Long> histogram = Histogram.valueOf(begin, end, elements);
 		Assert.assertEquals(histogram.length(), elements);
 		Assert.assertEquals(histogram.getHistogram(), new long[elements]);
 	}
-	
+
 	@Test
 	public void createInteger64() {
 		final Integer64 begin = Integer64.ZERO;
 		final Integer64 end = Integer64.valueOf(1000);
 		final int elements = 9;
-		
+
 		Histogram<Integer64> histogram = Histogram.valueOf(begin, end, elements);
 		Assert.assertEquals(histogram.length(), elements);
 		Assert.assertEquals(histogram.getHistogram(), new long[elements]);
 	}
-	
+
 	@Test
 	public void accumulate() {
 		final long begin = 0;
 		final long end = 10;
 		final int elements = 9;
-		
+
 		Histogram<Long> histogram = Histogram.valueOf(begin, end, elements);
 		for (int i = 0; i < elements*1000; ++i) {
 			histogram.accumulate(Long.valueOf(i%elements));
 		}
-		
+
 		final long[] expected = new long[9];
 		Arrays.fill(expected, 1000);
 		Assert.assertEquals(histogram.getHistogram(), expected);
 	}
-	
+
 	@Test
 	public void histogramIndex() {
 		final Random random = RandomRegistry.getRandom();
@@ -131,14 +132,14 @@ public class HistogramTest
 		for (int i = 0; i < parts.length; ++i) {
 			parts[i] = Double.valueOf(i);
 		}
-		
+
 		Histogram<Double> histogram = Histogram.valueOf(parts);
 		Double[] classes = histogram.getSeparators();
 		for (int i = 0; i < 1000; ++i) {
 			final Double value = random.nextDouble()*(parts.length + 1);
 			Assert.assertEquals(histogram.index(value), linearindex(classes, value));
 		}
-		
+
 		parts = new Double[]{1.0};
 		histogram = Histogram.valueOf(parts);
 		classes = histogram.getSeparators();
@@ -146,7 +147,7 @@ public class HistogramTest
 			final Double value = random.nextDouble()*(parts.length + 1);
 			Assert.assertEquals(histogram.index(value), linearindex(classes, value));
 		}
-		
+
 		parts = new Double[]{1.0, 2.0};
 		histogram = Histogram.valueOf(parts);
 		classes = histogram.getSeparators();
@@ -154,7 +155,7 @@ public class HistogramTest
 			final Double value = random.nextDouble()*(parts.length + 1);
 			Assert.assertEquals(histogram.index(value), linearindex(classes, value));
 		}
-		
+
 		parts = new Double[]{1.0, 2.0, 3.0};
 		histogram = Histogram.valueOf(parts);
 		classes = histogram.getSeparators();
@@ -163,7 +164,7 @@ public class HistogramTest
 			Assert.assertEquals(histogram.index(value), linearindex(classes, value));
 		}
 	}
-	
+
 	// The 'brute force' variante to test the binsearch one.
 	private static <C extends Comparable<C>> int linearindex(final C[] classes, final C value) {
 		int index = classes.length;
@@ -174,28 +175,28 @@ public class HistogramTest
 		}
 		return index;
 	}
-	
+
 	@Test(expectedExceptions = IllegalArgumentException.class)
 	public void histogramEmptyClasses() {
 		Histogram.valueOf(new Double[0]);
 	}
-	
+
 	@Test
 	public void histogram() {
 		final Random random = RandomRegistry.getRandom();
 		final Histogram<Double> histogram = Histogram.valueOf(new Double[]{1d, 2d, 3d, 4d, 5d});
-		
+
 		for (int i = 0; i < 600000; ++i) {
 			histogram.accumulate(random.nextDouble()*6);
 		}
 		Assert.assertEquals(histogram.getSamples(), 600000);
-		
+
 		final long[] hist = histogram.getHistogram();
 		for (int i = 0; i < hist.length; ++i) {
-			Assert.assertEquals((double)hist[i], 100000.0, 1000.0);
+			Assert.assertEquals(hist[i], 100000.0, 1000.0);
 		}
 	}
-	
+
 }
